requirements.txt
python-binance>=1.0.19
requests>=2.31.0
urllib3>=2.0.0
pillow>=10.0.0
pynacl>=1.5.0
pyinstaller>=6.0.0
